requirements.txt
openai-agents
python-dotenv
requests
feedparser
beautifulsoup4